Move core to peerDeps & constants to utils
