+++ packages/docusaurus-plugin-client-redirects/src/normalizePluginOptions.ts
@@ -7,7 +7,7 @@
 
 import {PluginOptions, RedirectOption, UserPluginOptions} from './types';
 import {Joi, PathnameSchema} from '@docusaurus/utils-validation';
-import {DEFAULT_PLUGIN_ID} from '@docusaurus/core/lib/constants';
+import {DEFAULT_PLUGIN_ID} from '@docusaurus/utils';
 
 export const DefaultPluginOptions: PluginOptions = {
   id: DEFAULT_PLUGIN_ID, // TODO temporary
