+++ packages/docusaurus-plugin-content-docs/package.json
@@ -18,7 +18,6 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/mdx-loader": "2.0.0-beta.9",
     "@docusaurus/utils": "2.0.0-beta.9",
     "@docusaurus/utils-validation": "2.0.0-beta.9",
@@ -47,6 +46,7 @@
     "utility-types": "^3.10.0"
   },
   "peerDependencies": {
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   },
