+++ packages/docusaurus-plugin-content-docs/src/__tests__/versions.test.ts
@@ -13,7 +13,7 @@ import {
   readVersionsMetadata,
 } from '../versions';
 import {DEFAULT_OPTIONS} from '../options';
-import {DEFAULT_PLUGIN_ID} from '@docusaurus/core/lib/constants';
+import {DEFAULT_PLUGIN_ID} from '@docusaurus/utils';
 import {PluginOptions, VersionMetadata} from '../types';
 import type {I18n} from '@docusaurus/types';
