+++ packages/docusaurus-plugin-debug/package.json
@@ -18,7 +18,6 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/utils": "2.0.0-beta.9",
     "fs-extra": "^10.0.0",
     "react-json-view": "^1.21.3",
@@ -28,6 +27,7 @@
     "@docusaurus/types": "2.0.0-beta.9"
   },
   "peerDependencies": {
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   },
