+++ packages/docusaurus-plugin-pwa/package.json
@@ -23,7 +23,6 @@
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.0",
     "@babel/plugin-proposal-optional-chaining": "^7.16.0",
     "@babel/preset-env": "^7.16.4",
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/theme-common": "2.0.0-beta.9",
     "@docusaurus/theme-translations": "2.0.0-beta.9",
     "@docusaurus/utils": "2.0.0-beta.9",
@@ -44,6 +43,7 @@
   },
   "peerDependencies": {
     "@babel/core": "^7.0.0",
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   }
