+++ packages/docusaurus-preset-classic/package.json
@@ -18,7 +18,6 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/plugin-content-blog": "2.0.0-beta.9",
     "@docusaurus/plugin-content-docs": "2.0.0-beta.9",
     "@docusaurus/plugin-content-pages": "2.0.0-beta.9",
@@ -30,6 +29,7 @@
     "@docusaurus/theme-search-algolia": "2.0.0-beta.9"
   },
   "peerDependencies": {
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   },
