+++ packages/docusaurus-theme-classic/package.json
@@ -23,7 +23,6 @@
     "update-code-translations": "node -e 'require(\"./update-code-translations.js\").run()'"
   },
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/plugin-content-blog": "2.0.0-beta.9",
     "@docusaurus/plugin-content-docs": "2.0.0-beta.9",
     "@docusaurus/plugin-content-pages": "2.0.0-beta.9",
@@ -56,6 +55,7 @@
     "utility-types": "^3.10.0"
   },
   "peerDependencies": {
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   },
