+++ packages/docusaurus-theme-live-codeblock/package.json
@@ -17,7 +17,6 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/theme-translations": "2.0.0-beta.9",
     "@docusaurus/utils-validation": "2.0.0-beta.9",
     "@docusaurus/utils": "2.0.0-beta.9",
@@ -33,6 +32,7 @@
     "@types/buble": "^0.20.1"
   },
   "peerDependencies": {
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   },
