+++ packages/docusaurus-theme-search-algolia/package.json
@@ -21,7 +21,6 @@
   },
   "dependencies": {
     "@docsearch/react": "^3.0.0-alpha.39",
-    "@docusaurus/core": "2.0.0-beta.9",
     "@docusaurus/theme-common": "2.0.0-beta.9",
     "@docusaurus/theme-translations": "2.0.0-beta.9",
     "@docusaurus/utils": "2.0.0-beta.9",
@@ -37,6 +36,7 @@
     "fs-extra": "^10.0.0"
   },
   "peerDependencies": {
+    "@docusaurus/core": "2.0.0-beta.9",
     "react": "^16.8.4 || ^17.0.0",
     "react-dom": "^16.8.4 || ^17.0.0"
   },
