+++ packages/docusaurus-theme-translations/package.json
@@ -26,7 +26,6 @@
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@docusaurus/core": "2.0.0-beta.9",
     "chalk": "^4.1.2",
     "lodash": "^4.17.20"
   },
