+++ packages/docusaurus-utils/src/constants.ts
@@ -28,6 +28,11 @@ export const GENERATED_FILES_DIR_NAME =
 
 export const SRC_DIR_NAME = 'src';
 export const STATIC_DIR_NAME = 'static';
+export const OUTPUT_STATIC_ASSETS_DIR_NAME = 'assets'; // files handled by webpack, hashed (can be cached aggressively)
 export const THEME_PATH = `${SRC_DIR_NAME}/theme`;
 export const DEFAULT_PORT = 3000;
 export const DEFAULT_PLUGIN_ID = 'default';
+
+// Temporary fix for https://github.com/facebook/docusaurus/issues/5493
+export const WEBPACK_URL_LOADER_LIMIT =
+  process.env.WEBPACK_URL_LOADER_LIMIT ?? 10000;
