+++ packages/docusaurus/src/server/themes/index.ts
@@ -7,7 +7,7 @@
 
 import {ThemeAliases, LoadedPlugin} from '@docusaurus/types';
 import path from 'path';
-import {THEME_PATH} from '../../constants';
+import {THEME_PATH} from '@docusaurus/utils';
 import themeAlias, {sortAliases} from './alias';
 
 const ThemeFallbackDir = path.resolve(__dirname, '../../client/theme-fallback');
