+++ packages/docusaurus/src/server/translations/__tests__/translationsExtractor.test.ts
@@ -14,7 +14,7 @@ import {
 import {getBabelOptions} from '../../../webpack/utils';
 import path from 'path';
 import {InitializedPlugin} from '@docusaurus/types';
-import {SRC_DIR_NAME} from '../../../constants';
+import {SRC_DIR_NAME} from '@docusaurus/utils';
 
 const TestBabelOptions = getBabelOptions({
   isServer: true,
